# backend/internal/api/openapi/core.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "KeepSave core management API",
    "version": "1.4.0-rc.1",
    "description": "Maintained identity, workspace and vault contract. Compatibility routes outside this contract are inventoried in docs/validation/2026-10-01-core-release/ACCEPTANCE.md. Definitions are not acceptance evidence; configuration and independent release reviews remain separate gates."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT human session"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
      "ErrorDetail": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "error_code": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "additionalProperties": false
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
      "Record": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "project_id": {
            "type": "string",
            "format": "uuid"
          },
          "environment": {
            "type": "string",
            "enum": [
              "alpha",
              "uat",
              "prod"
            ]
          },
          "key": {
            "type": "string"
          },
          "revision": {
            "type": "integer",
            "minimum": 1
          },
          "value": {
            "type": "string"
          },
          "deleted": {
            "type": "boolean"
          },
          "environment_id": {
            "type": "string",
            "format": "uuid"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "project_id",
          "environment",
          "key",
          "revision",
          "deleted",
          "environment_id",
          "created_at",
          "updated_at"
        ],
        "additionalProperties": false
      },
...
      "Bundle": {
        "type": "object",
        "properties": {
          "format": {
            "type": "string"
          },
          "ciphertext": {
            "type": "string",
            "format": "byte"
          },
          "nonce": {
            "type": "string",
            "format": "byte"
          },
          "sha256": {
            "type": "string",
            "minLength": 64,
            "maxLength": 64
          }
        },
        "required": [
          "format",
          "ciphertext",
          "nonce",
          "sha256"
        ],
        "additionalProperties": false
      },
      "Verification": {
        "type": "object",
        "properties": {
          "project_id": {
            "type": "string",
            "format": "uuid"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "entries": {
            "type": "integer",
            "minimum": 0
          },
          "revisions": {
            "type": "integer",
            "minimum": 0
          },
          "keys": {
            "type": "integer",
            "minimum": 0
          },
          "snapshots": {
            "type": "integer",
            "minimum": 0
          }
        },
        "required": [
          "project_id",
          "created_at",
          "entries",
          "revisions",
          "keys",
          "snapshots"
        ],
        "additionalProperties": false
      },
      "RestoreDiff": {
        "type": "object",
        "properties": {
          "secret_id": {
            "type": "string",
            "format": "uuid"
          },
          "environment": {
            "type": "string",
            "enum": [
              "alpha",
              "uat",
              "prod"
            ]
          },
          "key": {
            "type": "string"
          },
          "backup_revision": {
            "type": "integer",
            "minimum": 1
          },
          "current_revision": {
            "type": "integer",
            "minimum": 0
          },
          "status": {
            "type": "string"
          },
          "restorable": {
            "type": "boolean"
          }
        },
        "required": [
          "secret_id",
          "environment",
          "key",
          "backup_revision",
          "current_revision",
          "status",
          "restorable"
        ],
        "additionalProperties": false
      },
      "RestorePreview": {
        "type": "object",
        "properties": {
          "project_id": {
            "type": "string",
            "format": "uuid"
          },
          "backup_created_at": {
            "type": "string",
            "format": "date-time"
          },
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RestoreDiff"
            }
          }
        },
        "required": [
          "project_id",
          "backup_created_at",
          "records"
        ],
        "additionalProperties": false
      },
      "RestoreSelection": {
        "type": "object",
        "properties": {
          "secret_id": {
            "type": "string",
            "format": "uuid"
          },
          "backup_revision": {
            "type": "integer",
            "minimum": 1
          },
          "expected_current_revision": {
            "type": "integer",
            "minimum": 1
          }
        },
        "required": [
          "secret_id",
          "backup_revision",
          "expected_current_revision"
        ],
        "additionalProperties": false
      },
      "RestoreSelectedRequest": {
        "type": "object",
        "properties": {
          "bundle": {
            "$ref": "#/components/schemas/Bundle"
          },
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RestoreSelection"
            },
            "minItems": 1,
            "maxItems": 100
          }
        },
        "required": [
          "bundle",
          "records"
        ],
        "additionalProperties": true
      },
      "RestoredRecords": {
        "type": "object",
        "properties": {
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Record"
            }
          }
        },
        "required": [
          "records"
        ],
        "additionalProperties": false
      },
...
    }
  },
  "paths": {
...
    "/projects/{id}/backups/verify": {
      "post": {
        "operationId": "verifyEncryptedBackup",
        "summary": "verifyEncryptedBackup",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Verification"
                }
              }
            }
          },
          "400": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Bundle"
              }
            }
          }
        },
        "description": "PostgreSQL enrolled vault only. Recovery keys remain server-side. Selected revision checks apply; authority and deleted records are not restored."
      }
    },
    "/projects/{id}/backups/preview": {
      "post": {
        "operationId": "previewSelectedRestore",
        "summary": "previewSelectedRestore",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestorePreview"
                }
              }
            }
          },
          "400": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Bundle"
              }
            }
          }
        },
        "description": "PostgreSQL enrolled vault only. Recovery keys remain server-side. Selected revision checks apply; authority and deleted records are not restored."
      }
    },
    "/projects/{id}/backups/restore": {
      "post": {
        "operationId": "restoreSelectedRecords",
        "summary": "restoreSelectedRecords",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestoredRecords"
                }
              }
            }
          },
          "400": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestoreSelectedRequest"
              }
            }
          }
        },
        "description": "PostgreSQL enrolled vault only. Recovery keys remain server-side. Selected revision checks apply; authority and deleted records are not restored."
      }
    },
...
  }
}